+++ src/main/java/io/github/hapjava/server/impl/json/AccessoryController.java
@@ -18,6 +18,7 @@ import javax.json.Json;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
 import javax.json.JsonObjectBuilder;
+import javax.json.JsonWriter;
 
 public class AccessoryController {
 
@@ -64,9 +65,9 @@ public class AccessoryController {
               .add("services", serviceArrayBuilders.get(accessory.getId())));
     }
 
-    try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
-      Json.createWriter(baos)
-          .write(Json.createObjectBuilder().add("accessories", accessories).build());
+    try (ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        JsonWriter jsonWriter = Json.createWriter(baos)) {
+      jsonWriter.write(Json.createObjectBuilder().add("accessories", accessories).build());
       return new HapJsonResponse(baos.toByteArray());
     }
   }