+++ src/main/java/io/github/hapjava/server/impl/json/EventController.java
@@ -9,6 +9,7 @@ import javax.json.Json;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
 import javax.json.JsonObjectBuilder;
+import javax.json.JsonWriter;
 
 public class EventController {
 
@@ -24,8 +25,9 @@ public class EventController {
 
     JsonObject data = Json.createObjectBuilder().add("characteristics", characteristics).build();
 
-    try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
-      Json.createWriter(baos).write(data);
+    try (ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        JsonWriter jsonWriter = Json.createWriter(baos)) {
+      jsonWriter.write(data);
       byte[] dataBytes = baos.toByteArray();
 
       return new EventResponse(dataBytes);
@@ -44,9 +46,9 @@ public class EventController {
     }
 
     JsonObject data = Json.createObjectBuilder().add("characteristics", characteristics).build();
-
-    try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
-      Json.createWriter(baos).write(data);
+    try (ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        JsonWriter jsonWriter = Json.createWriter(baos)) {
+      jsonWriter.write(data);
       byte[] dataBytes = baos.toByteArray();
 
       return new EventResponse(dataBytes);