+++ src/main/java/io/github/hapjava/characteristics/impl/base/FloatCharacteristic.java
@@ -130,7 +130,7 @@ public abstract class FloatCharacteristic extends BaseCharacteristic<Double> {
 
   @Override
   protected void setValue(Double value) throws Exception {
-    setter.get().accept(value);
+    if (setter.isPresent()) setter.get().accept(value);
   }
 
   /** {@inheritDoc} */