+++ CHANGES.md
@@ -1,6 +1,7 @@
 # HAP-Java 2.0.1
 ## Fixes
 * Log accessory names instead of futures. [#150](https://github.com/hap-java/HAP-Java/issues/150)
+* Fix rotation speed data type (BREAKING API CHANGE). According to HAP specification it must be float
 
 # HAP-Java 2.0.0
 * major refactoring to support optional characteristics